fix : List - move AddSampleData EditCurrentPrice folder in StockTableMenu folder
