+++ src/views/List/StockListContent/StockTableMenu/AddSampleData/index.tsx
@@ -1,13 +1,13 @@
-import { BorderButton } from '../../../components/Button';
+import { memo } from 'react';
+import styled from 'styled-components';
 import { useDispatch, useSelector } from 'react-redux';
 
-import { selectIsLoggedIn } from '../../../features/user/selectors';
-import userDataService from '../../../service/userData/userData';
 import { addStockSampleData } from '@/features';
 import { useShowAddSampleBtn } from './useShowAddSampleBtn';
 import Icon from '@/components/Icon';
-import styled from 'styled-components';
-import { memo } from 'react';
+import { BorderButton } from '@/components/Button';
+import userDataService from '@/service/userData/userData';
+import { selectIsLoggedIn } from '@/features/user/selectors';
 
 const AddSampleData = () => {
   const dispatch = useDispatch();
