+++ src/views/List/StockListContent/StockTableMenu/EditCurrentPrice/EditCurrentPriceModal.tsx
@@ -12,21 +12,25 @@ import {
   TableHeader,
   TableRow,
 } from '@/components/Table';
-import { HeaderItemProps } from '../Header/HeaderItem';
+
 import {
   selectStockInfoById,
   selectStocks,
 } from '@/features/stockList/selectors';
-import { InputCell } from '../StockItem/components';
+
 import { selectIsLoggedIn } from '@/features/user/selectors';
 import { updateStocksCurrentPrice } from '@/features/stockList/stockListSlice';
 import userStocksService from '@/service/userStocks/userStocks';
-import { PurchasedInputChangeProps } from '../StockItem/PurchasedStock/PurchasedStock';
+
 import { getFixedLocaleString } from '@/utils';
-import { checkCurrentPrice } from '../StockItem/validity';
+
 import { StockMainInfo } from '@/features/stockList/type';
 import { StyledModalContainer, StyledModalMessage } from '@/components/Modal';
 import Flex from '@/components/Flex';
+import { PurchasedInputChangeProps } from '@/views/List/StockItem/PurchasedStock/PurchasedStock';
+import { HeaderItemProps } from '@/views/List/Header/HeaderItem';
+import { checkCurrentPrice } from '@/views/List/StockItem/validity';
+import { InputCell } from '@/views/List/StockItem/components';
 
 export interface CurrentPriceChanges {
   [key: string]: StockMainInfo['currentPrice'];
